@@ -3,6 +3,7 @@
   "private": true,
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "scripts": {
     "dev": "vite serve",
     "debug": "node --inspect ./node_modules/vite/bin/vite serve",
@@ -15,20 +16,18 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@types/react": "^18.0.26",
+    "@vitejs/plugin-react": "^3.0.0",
     "gh-pages": "^4.0.0",
     "rimraf": "^3.0.2",
-    "serve": "^14.0.1",
-    "vite": "^3.1.0",
+    "serve": "^14.1.2",
+    "vite": "^4.0.1",
     "vite-pages-theme-doc": "workspace:*",
-    "vite-plugin-mdx": "^3.5.11",
     "vite-plugin-react-pages": "workspace:*"
   }
 }
