@@ -8,53 +8,48 @@
     "build": "pnpm -r --filter='./packages/*' run build",
     "dev": "pnpm -r --parallel --filter='./packages/*' run dev",
     "format": "prettier --write ./**/*.{js,ts,tsx}",
-    "ver": "pnpm recursive --filter create-vite-pages --filter vite-plugin-react-pages --filter vite-pages-theme-basic --filter vite-pages-theme-doc exec -- npm version",
+    "ver": "pnpm recursive --filter create-vite-pages --filter vite-plugin-react-pages --filter vite-pages-theme-doc exec -- npm version",
     "ver:p": "npm run ver -- patch",
     "ver:mi": "npm run ver -- minor",
     "ver:major": "npm run ver -- major",
-    "pub": "pnpm -r publish --filter create-vite-pages --filter vite-plugin-react-pages --filter vite-pages-theme-basic --filter vite-pages-theme-doc",
+    "ver:alpha": "npm run ver -- prerelease --preid alpha",
+    "pub": "pnpm -r publish --filter create-vite-pages --filter vite-plugin-react-pages --filter vite-pages-theme-doc",
     "update-deps": "pnpm -r update -i -L",
-    "update-local-deps": "pnpm update -r vite-plugin-react-pages vite-pages-theme-basic vite-pages-theme-doc --workspace --filter \"!./packages/create-project/**\"",
+    "update-local-deps": "pnpm update -r vite-plugin-react-pages vite-pages-theme-doc --workspace --filter \"!./packages/create-project/**\"",
     "test": "playwright test -c ./test-setup",
     "test-serve": "cross-env VITE_PAGES_MODE=serve pnpm test",
     "test-build": "cross-env VITE_PAGES_MODE=build pnpm test",
     "test-ssr": "cross-env VITE_PAGES_MODE=ssr pnpm test",
     "test-debug": "pnpm test -- --debug",
     "build:docs": "pnpm --filter doc-site run ssr-for-deploy",
-    "postinstall": "node ./scripts/linkBins.js",
     "install-test-deps": "playwright install --with-deps chromium",
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@playwright/test": "~1.23.4",
-    "@pnpm/link-bins": "^7.2.4",
-    "@pnpm/logger": "^4.0.0",
-    "@types/node": "^17.0.42",
+    "@playwright/test": "~1.29.0",
+    "@pnpm/link-bins": "^8.0.4",
+    "@pnpm/logger": "^5.0.0",
+    "@types/node": "^18.11.16",
     "@types/wait-on": "^5.3.1",
     "cross-env": "^7.0.3",
     "execa": "^5.1.1",
     "get-port": "^5.1.1",
-    "husky": "^8.0.1",
-    "prettier": "^2.7.1",
+    "husky": "^8.0.2",
+    "prettier": "^2.8.1",
     "sirv": "^2.0.2",
     "slash": "^3.0.0",
-    "typescript": "^4.8.3",
+    "typescript": "^4.9.4",
     "wait-on": "^6.0.1"
   },
   "pnpm": {
     "overrides": {
-      "react": "^17.0.1",
-      "react-dom": "^17.0.1",
       "vite-plugin-react-pages": "workspace:*",
-      "vite-pages-theme-doc": "workspace:*",
-      "vite-pages-theme-basic": "workspace:*"
+      "vite-pages-theme-doc": "workspace:*"
     },
     "peerDependencyRules": {
       "ignoreMissing": [
         "react",
         "react-dom",
-        "@mdx-js/mdx",
-        "@alifd/meet-react",
         "typescript",
         "rollup",
         "@babel/core"
