@@ -2,6 +2,7 @@
   "name": "app-demo",
   "version": "1.0.2",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite serve",
     "build": "rimraf dist && vite build --outDir dist && serve -s dist",
@@ -11,22 +12,19 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0"
   },
   "devDependencies": {
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@types/node": "^17.0.42",
-    "@types/react": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@mdx-js/react": "^2.2.1",
+    "@types/node": "^18.11.16",
+    "@types/react": "^18.0.26",
+    "@vitejs/plugin-react": "^3.0.0",
     "rimraf": "^3.0.2",
-    "serve": "^14.0.1",
-    "vite": "^3.1.0",
-    "vite-pages-theme-doc": "^3.1.4",
-    "vite-plugin-mdx": "^3.5.11",
-    "vite-plugin-react-pages": "^3.1.8"
+    "serve": "^14.1.2",
+    "vite": "^4.0.1",
+    "vite-pages-theme-doc": "^4.0.0-alpha.1",
+    "vite-plugin-react-pages": "^4.0.0-alpha.2"
   }
 }
