@@ -9,9 +9,9 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/react": "^17.0.3",
-    "react": "^17.0.1",
+    "@types/react": "^18.0.26",
+    "react": "^18.2.0",
     "rimraf": "^3.0.2",
-    "typescript": "^4.8.3"
+    "typescript": "^4.9.4"
   }
 }
