@@ -12,19 +12,18 @@
     "debug:ssr": "rimraf dist && node --inspect-brk ./node_modules/vite-plugin-react-pages/bin/vite-pages.js ssr"
   },
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0"
   },
   "devDependencies": {
-    "@mdx-js/mdx": "^1.6.22",
-    "@types/react": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@mdx-js/mdx": "^2.2.1",
+    "@types/react": "^18.0.26",
+    "@vitejs/plugin-react": "^3.0.0",
     "rimraf": "^3.0.2",
-    "serve": "^14.0.1",
-    "vite": "^3.1.0",
-    "vite-plugin-mdx": "^3.5.11",
+    "sass": "^1.57.0",
+    "serve": "^14.1.2",
+    "vite": "^4.0.1",
     "vite-plugin-react-pages": "workspace:*"
   }
 }
