@@ -1,8 +1,8 @@
 import React from 'react'
-import { Redirect } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 
 const Component404 = () => {
-  return <Redirect to="/" />
+  return <Navigate to="/" replace />
 }
 
 export default Component404
