@@ -12,7 +12,7 @@ export default createTheme({
       activeIfMatch: {
         // match all first-level paths
         path: '/:foo',
-        exact: true,
+        end: true,
       },
     },
     {
