@@ -1,11 +1,11 @@
-import { createTheme } from 'vite-pages-theme-basic'
+import { createTheme } from 'vite-pages-theme-doc'
 
 export default createTheme({
   topNavs: [
-    { text: 'index', path: '/' },
-    { text: 'Vite', href: 'https://github.com/vitejs/vite' },
+    { label: 'index', path: '/' },
+    { label: 'Vite', href: 'https://github.com/vitejs/vite' },
     {
-      text: 'Vite Pages',
+      label: 'Vite Pages',
       href: 'https://github.com/vitejs/vite-plugin-react-pages',
     },
   ],
