@@ -1,4 +1,5 @@
-// import the dev entry
+// import the src directly instead of build output
 // useful for local development in this repo
 // actual users don't need to do this, they should import from 'vite-pages-theme-doc'
-export * from 'vite-pages-theme-doc/src/index.dev'
+// export * from 'vite-pages-theme-doc/src/index'
+export * from 'vite-pages-theme-doc'
