@@ -1,12 +1,7 @@
 import { defineConfig } from 'vite'
 import * as path from 'path'
 import react from '@vitejs/plugin-react'
-import _mdx from 'vite-plugin-mdx'
 import pages from 'vite-plugin-react-pages'
-import { LessPluginRemoveAntdGlobalStyles } from 'less-plugin-remove-antd-global-styles'
-
-// vite-plugin-mdx is not esm friendly currently
-const mdx = ((_mdx as any).default || _mdx) as typeof _mdx
 
 export default defineConfig({
   resolve: {
@@ -16,22 +11,13 @@ export default defineConfig({
   },
   plugins: [
     react(),
-    mdx(),
     pages({
       pagesDir: path.join(__dirname, 'pages'),
+      // useHashRouter: true
     }),
   ],
   // theme local dev
   css: {
-    preprocessorOptions: {
-      less: {
-        modifyVars: {
-          'ant-prefix': 'vp-antd',
-        },
-        javascriptEnabled: true,
-        plugins: [new LessPluginRemoveAntdGlobalStyles()],
-      },
-    },
     modules: {
       generateScopedName: `vp-local-[local]`,
     },
