@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../dist/node/cli')
+import '../dist/node-esm/cli.mjs'
