@@ -1,14 +1,20 @@
-import { useMemo, createContext } from 'react'
+import { useMemo, useEffect, useState } from 'react'
 import { dequal } from 'dequal'
-import type { SetAtom } from 'jotai/core/types'
-import { atom, useAtom } from './jotai'
-import { atomFamily, useAtomValue, useUpdateAtom } from './jotai/utils'
-import type { PageLoaded, UseStaticData, Theme } from '../../clientTypes'
+import type { SetAtom } from 'jotai/core/atom'
+import { atom, useAtom } from 'jotai'
+import { atomFamily, useAtomValue, useUpdateAtom } from 'jotai/utils'
+import type {
+  PageLoaded,
+  UseStaticData,
+  Theme,
+  UseAllPagesOutlines,
+} from '../../clientTypes'
 
 export let useTheme: () => Theme
 export let usePagePaths: () => string[]
 export let usePageModule: (path: string) => Promise<PageModule> | undefined
 export let useStaticData: UseStaticData
+export let useAllPagesOutlines: UseAllPagesOutlines
 
 interface PageModule {
   ['default']: PageLoaded
@@ -27,8 +33,9 @@ const initialPagePaths = Object.keys(initialPages)
 // by the same Provider. It also mutates during render, which is
 // generally discouraged, but in this case it's okay.
 if (import.meta.hot) {
-  let setTheme: SetAtom<{ Theme: Theme }> | undefined
+  let setTheme: SetAtom<{ Theme: Theme }, void> | undefined
   import.meta.hot!.accept('/@react-pages/theme', (module) => {
+    // console.log('@@hot update /@react-pages/theme', module)
     if (!module) {
       console.error('unexpected hot module', module)
       return
@@ -43,8 +50,9 @@ if (import.meta.hot) {
     return Theme
   }
 
-  let setPages: SetAtom<any> | undefined
+  let setPages: SetAtom<any, void> | undefined
   import.meta.hot!.accept('/@react-pages/pages', (module) => {
+    // console.log('@@hot update /@react-pages/pages', module)
     if (!module) {
       console.error('unexpected hot module', module)
       return
@@ -52,9 +60,20 @@ if (import.meta.hot) {
     setPages?.(module.default)
   })
 
+  let setAllPagesOutlines: SetAtom<any, void> | undefined
+  import.meta.hot!.accept('/@react-pages/allPagesOutlines', (module) => {
+    // console.log('@@hot update /@react-pages/allPagesOutlines', module)
+    if (!module) {
+      console.error('unexpected hot module', module)
+      return
+    }
+    setAllPagesOutlines?.(module)
+  })
+
   const pagesAtom = atom(initialPages)
   const pagePathsAtom = atom(initialPagePaths.sort())
   const staticDataAtom = atom(toStaticData(initialPages))
+  const allPagesOutlinesAtom = atom(initialPages)
 
   const setPagesAtom = atom(null, (get, set, newPages: any) => {
     let newStaticData: Record<string, any> | undefined
@@ -106,16 +125,20 @@ if (import.meta.hot) {
     }
   })
 
-  const dataAtoms = atomFamily((path: string) => (get) => {
-    const pages = get(pagesAtom)
-    return pages[path]
-  })
+  const dataAtoms = atomFamily((path: string) =>
+    atom((get) => {
+      const pages = get(pagesAtom)
+      return pages[path]
+    })
+  )
 
-  const staticDataAtoms = atomFamily((path: string) => (get) => {
-    const pages = get(pagesAtom)
-    const page = pages[path]
-    return page?.staticData
-  })
+  const staticDataAtoms = atomFamily((path: string) =>
+    atom((get) => {
+      const pages = get(pagesAtom)
+      const page = pages[path]
+      return page?.staticData
+    })
+  )
 
   usePagePaths = () => {
     setPages = useUpdateAtom(setPagesAtom)
@@ -138,6 +161,19 @@ if (import.meta.hot) {
     }
     return useAtomValue(staticData)
   }
+
+  useAllPagesOutlines = (timeout: number) => {
+    const [data, set] = useAtom(allPagesOutlinesAtom)
+    setAllPagesOutlines = set
+    useEffect(() => {
+      setTimeout(() => {
+        import('/@react-pages/allPagesOutlines').then((mod) => {
+          set(mod)
+        })
+      }, timeout)
+    }, [])
+    return data
+  }
 }
 
 // Static mode
@@ -156,6 +192,17 @@ else {
     }
     return toStaticData(initialPages)
   }
+  useAllPagesOutlines = (timeout: number) => {
+    const [data, set] = useState<any>()
+    useEffect(() => {
+      setTimeout(() => {
+        import('/@react-pages/allPagesOutlines').then((mod) => {
+          set(mod)
+        })
+      }, timeout)
+    }, [])
+    return data
+  }
 }
 
 function toStaticData(pages: Record<string, any>) {
@@ -168,9 +215,11 @@ function toStaticData(pages: Record<string, any>) {
 
 if ((globalThis as any)['__vite_pages_use_static_data']) {
   throw new Error(
-    `[vite-pages] useStaticData already exists on window. There are multiple vite-pages apps in this page. Please report this to vite-pages.`
+    `[vite-pages] global hooks (.e.g useStaticData) already exists on window. It means there are multiple vite-pages runtime in this page. Please report this to vite-pages.`
   )
 } else {
-  // make it available to vite-plugin-react-pages/client
+  // make them available in vite-plugin-react-pages/client
   ;(globalThis as any)['__vite_pages_use_static_data'] = useStaticData
+  ;(globalThis as any)['__vite_pages_use_all_pages_outlines'] =
+    useAllPagesOutlines
 }
