@@ -1,4 +1,10 @@
+/// <reference types="remark-mdx" />
 import type { Root } from 'mdast'
+import type { MdxjsEsm } from 'mdast-util-mdx'
+import {
+  createJSXWithSpreadPropsNode,
+  createNameSpaceImportNode,
+} from '../../utils/mdastUtils'
 
 export function DemoMdxPlugin() {
   return transformer
@@ -6,29 +12,40 @@ export function DemoMdxPlugin() {
   function transformer(tree: Root, file: any) {
     const children = tree.children
 
-    const addImports: string[] = []
+    const addImports: MdxjsEsm[] = []
 
-    children.forEach((child: any) => {
-      if ((child.type as string) === 'jsx') {
-        const regexp = /<Demo\s+src=["'](.*?)["']/
-        const match = (child.value as string).match(regexp)
-        if (match) {
-          const src = match[1]
+    children.forEach((child, index) => {
+      // find jsx node like: <Demo src="srcValue" />
+      if (child.type === 'mdxJsxFlowElement' && child.name === 'Demo') {
+        const srcAttr = child.attributes.find(
+          (attr) =>
+            attr.type === 'mdxJsxAttribute' &&
+            attr.name === 'src' &&
+            typeof attr.value === 'string'
+        )
+        const srcValue = srcAttr?.value
+        if (srcValue) {
           const nextIndex = addImports.length
           const varName = `_demo${nextIndex}`
-          addImports.push(`import * as ${varName} from "${src}?demo";`)
-          child.value = `<Demo {...${varName}} />`
+          // add import:
+          // import * as varName from "${srcValue}?tsInfo=${nameValue}"
+          addImports.push(
+            createNameSpaceImportNode({
+              name: varName,
+              from: `${srcValue}?demo`,
+            })
+          )
+          // replace this node with:
+          // <Demo {...varName} />
+          children.splice(
+            index,
+            1,
+            createJSXWithSpreadPropsNode({ Component: 'Demo', props: varName })
+          )
         }
       }
     })
 
-    children.unshift(
-      ...addImports.map((importStr) => {
-        return {
-          type: 'import',
-          value: importStr,
-        } as any
-      })
-    )
+    children.unshift(...addImports)
   }
 }
