@@ -1,4 +1,10 @@
+/// <reference types="remark-mdx" />
 import type { Root } from 'mdast'
+import type { MdxjsEsm } from 'mdast-util-mdx'
+import {
+  createJSXWithSpreadPropsNode,
+  createNameSpaceImportNode,
+} from '../../utils/mdastUtils'
 
 export function TsInfoMdxPlugin() {
   return transformer
@@ -6,32 +12,52 @@ export function TsInfoMdxPlugin() {
   function transformer(tree: Root, file: any) {
     const children = tree.children
 
-    const addImports: string[] = []
+    const addImports: MdxjsEsm[] = []
 
-    children.forEach((child: any) => {
-      if ((child.type as string) === 'jsx') {
-        const regexp = /<TsInfo\s+src=["'](.*?)["']\s+name=["'](.*?)["']/
-        const match = (child.value as string).match(regexp)
-        if (match) {
-          const src = match[1]
-          const name = match[2]
+    children.forEach((child, index) => {
+      // find jsx node like: <TsInfo src="srcValue" name="nameValue" />
+      if (child.type === 'mdxJsxFlowElement' && child.name === 'TsInfo') {
+        const srcAttr = child.attributes.find(
+          (attr) =>
+            attr.type === 'mdxJsxAttribute' &&
+            attr.name === 'src' &&
+            typeof attr.value === 'string'
+        )
+        const srcValue = srcAttr?.value
+
+        const nameAttr = child.attributes.find(
+          (attr) =>
+            attr.type === 'mdxJsxAttribute' &&
+            attr.name === 'name' &&
+            typeof attr.value === 'string'
+        )
+        const nameValue = nameAttr?.value
+
+        if (srcValue && nameValue) {
           const nextIndex = addImports.length
           const varName = `_tsInfo${nextIndex}`
+          // add import:
+          // import * as varName from "${srcValue}?tsInfo=${nameValue}"
           addImports.push(
-            `import * as ${varName} from "${src}?tsInfo=${name}";`
+            createNameSpaceImportNode({
+              name: varName,
+              from: `${srcValue}?tsInfo=${nameValue}`,
+            })
           )
-          child.value = `<TsInfo {...${varName}} />`
+          // replace this node with:
+          // <TsInfo {...varName} />
+          children.splice(
+            index,
+            1,
+            createJSXWithSpreadPropsNode({
+              Component: 'TsInfo',
+              props: varName,
+            })
+          )
         }
       }
     })
 
-    children.unshift(
-      ...addImports.map((importStr) => {
-        return {
-          type: 'import',
-          value: importStr,
-        } as any
-      })
-    )
+    children.unshift(...addImports)
   }
 }
