@@ -1,15 +1,16 @@
 import React from 'react'
-import Highlight, { defaultProps } from 'prism-react-renderer'
-import theme from 'prism-react-renderer/themes/github'
+import * as _cjsExports from 'prism-react-renderer'
+const { default: Highlight, defaultProps } = commonjsExportsInterop(_cjsExports)
+import theme from 'prism-react-renderer/themes/github/index.cjs'
 import type { Language } from 'prism-react-renderer'
 import { useCopyToClipBoard } from './useCopyToClipBoard'
 import s from './CodeBlock.module.less'
+import { commonjsExportsInterop } from '../../utils'
 
 // copied from https://mdxjs.com/guides/syntax-highlighting
 
 interface Props {
-  readonly className?: `language-${Language}` | ''
-  readonly children?: string
+  readonly className?: `language-${Language}` | '' | string
   readonly style?: React.CSSProperties
 }
 
@@ -17,11 +18,14 @@ const CodeBlock = ({
   children = '',
   className = '',
   style: propStyle,
-}: Props) => {
+}: React.PropsWithChildren<Props>) => {
   // with ```language\n``` md syntax, mdx will pass language in className
   const language = className?.replace(/language-/, '') as Language
 
   const { hasCopied, copyToClipBoard } = useCopyToClipBoard()
+
+  if (typeof children !== 'string')
+    throw new Error('unexpected children of CodeBlock: should be string')
 
   return (
     <Highlight
